CN-B-9: Fix chat logic and properties name in JSON

+++ src/main/java/ru/sirius/concordia/core/model/dto/data/CountDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @Builder
 public class CountDTO {
-    private final int count;
+    private final Long count;
 }
